[REFACTOR] Separate global logging setup and logger itself
