+++ pandora-common/src/gameLogic/characterModifiers/characterModifiersSubsystemServer.ts
@@ -4,7 +4,8 @@ import type { AssetManager } from '../../assets/assetManager.ts';
 import { GetRestrictionOverrideConfig } from '../../assets/state/characterStateTypes.ts';
 import type { AssetFrameworkGlobalState } from '../../assets/state/globalState.ts';
 import { LIMIT_CHARACTER_MODIFIER_INSTANCE_COUNT } from '../../inputLimits.ts';
-import { GetLogger, Logger } from '../../logging.ts';
+import type { Logger } from '../../logging/logger.ts';
+import { GetLogger } from '../../logging/logging.ts';
 import type { CurrentSpaceInfo } from '../../space/index.ts';
 import { AssertNever, AssertNotNullable } from '../../utility/misc.ts';
 import { ArrayIncludesGuard } from '../../validation.ts';
