+++ pandora-common/test/networking/connection.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 import { EmitterWithAck, IEmpty, MESSAGE_HANDLER_DEBUG_MESSAGES } from '../../src/index.ts';
-import { GetLogger, LogLevel } from '../../src/logging.ts';
+import { GetLogger, LogLevel } from '../../src/logging/index.ts';
 import { ConnectionBase } from '../../src/networking/connection.ts';
 
 const mockEmitCB = jest.fn((_event: string, _arg: unknown, _cb?: (arg: unknown) => void) => {/**nothing */ });
